pyproject.toml
[tool.poetry]
name = "aiohttp_learning"
version = "0.1.0"
description = ""
authors = ["artur <shadaraevd@gmail.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
aiohttp-apispec = "^2.2.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "app.web.app:run_app"